Fix testsg

+++ tests/utils/file_test.cpp
@@ -1,5 +1,5 @@
+#include "yarl/test.hpp"
 #include "yarl/utils/file.hpp"
-#include "yarl/yarl_test.hpp"
 
 namespace yarl {
 
@@ -12,7 +12,7 @@ TEST(Utils, path_split) {
   std::vector<std::string> splits;
 
   splits = path_split("/a/b/c.yaml");
-  ASSERT_EQ(3, splits.size());
+  ASSERT_EQ(3, (int) splits.size());
   ASSERT_EQ("a", splits[0]);
   ASSERT_EQ("b", splits[1]);
   ASSERT_EQ("c.yaml", splits[2]);
